refactor cast suites
